Add files via upload
DermAid Skin Disease Detection

This repository contains the machine learning model and codebase for DermAid, an image-based skin disease detection application developed for the Software Development Group Project. The application focuses on the detection of 19 different skin diseases using convolutional neural networks (CNNs) and the ResNet50 architecture.

Key Features:

#Implementation of a CNN-based skin disease detection model using TensorFlow and Keras.
#Integration of the ResNet50 architecture for feature extraction.
#Data preprocessing techniques including image resizing, normalization, and augmentation.
#Model training, evaluation, and validation procedures.
#Exporting the trained model to TensorFlow Lite format for deployment on mobile devices.
#Creation of a classification report and visualization of model performance using confusion matrices and classification reports.

The codebase provides a comprehensive approach to building and evaluating a machine learning model for skin disease detection, serving as a foundation for the DermAid application.